mobile_app/backend/requirements.txt
albucore==0.0.13
albumentations==1.4.10
altgraph==0.17.4
annotated-types==0.7.0
anyio==4.6.2.post1
asgiref==3.8.1
astor==0.8.1
astroid==3.3.5
asttokens==3.0.0
autopep8==2.3.1
beautifulsoup4==4.12.3
black==24.10.0
blinker==1.8.2
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
coloredlogs==15.0.1
comm==0.2.2
contourpy==1.3.0
cycler==0.12.1
Cython==3.0.11
debugpy==1.8.9
decorator==5.1.1
dill==0.3.9
Django==3.1.12
djangorestframework==3.11.2
djongo==1.3.7
elastic-transport==8.15.0
elasticsearch==8.15.1
executing==2.1.0
filelock==3.16.1
fire==0.7.0
Flask==3.0.3
flatbuffers==24.3.25
fonttools==4.53.1
fsspec==2024.10.0
gitdb==4.0.11
GitPython==3.1.43
h11==0.14.0
httpcore==1.0.7
httpx==0.28.0
humanfriendly==10.0
idna==3.10
imageio==2.36.1
imgaug==0.4.0
ipykernel==6.29.5
ipython==8.30.0
isort==5.13.2
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.4
joblib==1.4.2
jupyter_client==8.6.3
jupyter_core==5.7.2
kiwisolver==1.4.7
lazy_loader==0.4
lmdb==1.5.1
lxml==5.3.0
MarkupSafe==2.1.5
matplotlib==3.9.2
matplotlib-inline==0.1.7
mccabe==0.7.0
mpmath==1.3.0
mypy-extensions==1.0.0
nest-asyncio==1.6.0
networkx==3.4.2
nltk==3.9.1
numpy==1.26.4
onnx==1.17.0
onnxruntime==1.20.1
onnxslim==0.1.42
opencv-contrib-python==4.10.0.84
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
opt-einsum==3.3.0
packaging==24.2
paddleocr==2.9.1
paddlepaddle==2.6.2
pandas==2.2.2
parso==0.8.4
pathspec==0.12.1
pefile==2024.8.26
pillow==10.4.0
platformdirs==4.3.6
plotly==5.24.1
plyer==2.1.0
prompt_toolkit==3.0.48
protobuf==3.20.2
psutil==6.1.0
pure_eval==0.2.3
py-cpuinfo==9.0.0
pyclipper==1.3.0.post6
pycodestyle==2.12.1
pydantic==2.10.2
pydantic_core==2.27.1
pyg_lib==0.4.0+pt25cpu
Pygments==2.18.0
pyinstaller==6.10.0
pyinstaller-hooks-contrib==2024.8
pylint==3.3.1
pymongo==3.11.4
pyparsing==3.1.4
pypiwin32==223
PyQt5==5.15.11
PyQt5-Qt5==5.15.2
PyQt5_sip==12.15.0
pyreadline3==3.5.4
pystray==0.19.5
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-docx==1.1.2
pytz==2024.2
pywin32==308
pywin32-ctypes==0.2.3
PyYAML==6.0.2
pyzmq==26.2.0
RapidFuzz==3.10.1
regex==2024.11.6
requests==2.32.3
scikit-image==0.24.0
scikit-learn==1.5.2
scipy==1.14.1
seaborn==0.13.2
setuptools==75.6.0
shapely==2.0.6
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.6
sqlparse==0.2.4
stack-data==0.6.3
sympy==1.13.1
tenacity==9.0.0
termcolor==2.5.0
thop==0.1.1.post2209072238
threadpoolctl==3.5.0
tifffile==2024.9.20
tomli==2.2.1
tomlkit==0.13.2
torch==2.2.0
torch_cluster==1.6.3+pt25cpu
torch_scatter==2.1.2+pt25cpu
torch_sparse==0.6.18+pt25cpu
torch_spline_conv==1.2.2+pt25cpu
torchaudio==2.2.0
torchvision==0.17.0
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2024.1
ultralytics==8.3.39
ultralytics-thop==2.0.12
urllib3==2.2.3
wcwidth==0.2.13
Werkzeug==3.0.4
wheel==0.45.1
win10toast==0.9
